feat: complete SessionSummary to SessionResult unification with full implementation
Major changes:
- Extended models::SessionResult with all necessary fields for backward compatibility:
  * Added total_session_time (alias for session_duration)
  * Added total_challenges_completed/attempted (aliases for stages_completed/attempted)
  * Added total_skips_used (alias for stages_skipped)
  * Added partial effort tracking fields
  * Added best/worst performance tracking fields

- Implemented all required methods in SessionResult:
  * add_stage_result() for tracking completed stages
  * add_skip() for handling skipped stages
  * add_partial_effort() for tracking incomplete attempts
  * total_effort_keystrokes() and total_effort_mistakes() helper methods
  * finalize_session() for calculating final metrics
  * get_session_completion_status() for summary text

- Updated session_tracker.rs to use models::SessionResult instead of duplicate struct
- Updated all imports and exports throughout codebase
- Fixed missing metrics module error in scoring/mod.rs

All functionality preserved while achieving complete terminology unification.
No breaking changes to existing APIs.
