--- src/scoring/metrics.rs
@@ -1,2 +1,0 @@
-// Re-export StageResult as TypingMetrics for backward compatibility
-pub use crate::models::StageResult as TypingMetrics;
