+++ src/scoring/mod.rs
@@ -1,6 +1,4 @@
 pub mod engine;
-pub mod metrics;
 
 pub use engine::ScoringEngine;
-pub use metrics::TypingMetrics;
 pub use crate::models::{StageResult, RankTier, Rank};
